testgenerator/requirements.txt
gigachat>=0.1.0
PyPDF2>=3.0.0
langchain>=0.1.0
langchain-community>=0.0.10
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0 